Store money as Decimal instead of Float
- products.price, orders.total_amount and order_items.price_at_purchase
  are now Numeric(10, 2) instead of Float; migration 739d351e6fb5
  converts the existing columns.
- Schemas use Decimal. ProductCreate.price must be > 0 with at most
  2 decimal places.
- Cart and order totals are summed as Decimal.
- The API sends money as a string (e.g. "19.99") everywhere: product,
  order and cart responses and the Redis product cache.

+++ app/models/cartItem.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.database.base import Base
@@ -13,7 +13,7 @@ class OrderItem(Base):
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
 
     quantity = Column(Integer, nullable=False)
-    price_at_purchase = Column(Float, nullable=False)
+    price_at_purchase = Column(Numeric(10, 2), nullable=False)
 
     order = relationship(
         "Order",
